Fixed Comments for HPStationTransformUtil
