+++ src/main/java/frc/robot/Constants.java
@@ -23,6 +23,9 @@ public final class Constants {
   public static final double PLACEHOLDER_DOUBLE = 0.0;
   public static final boolean PLACEHOLDER_BOOLEAN = false;
 
+  public static final double HP_INTAKE_TAG_LOCATION_OFFSET_X_INCHES = 10.6;
+  public static final double HP_INTAKE_TAG_LOCATION_OFFSET_Y_INCHES = 14.6;
+
   public static enum limelightLedMode {
     PIPELINE_MODE,
     OFF,
